+++ lib/screens/widgets/gradient_button.dart
@@ -1,0 +1,30 @@
+// import 'package:flutter/material.dart';
+
+// class GradientButton extends StatelessWidget {
+//   // final String title;
+//   // final String 
+//   const GradientButton({super.key});
+
+//   @override
+//   Widget build(BuildContext context) {
+//     return Padding(
+//         padding: const EdgeInsets.all(8),
+//         child: ElevatedButton(
+//             style: ButtonStyle(
+//               shape: MaterialStatePropertyAll(RoundedRectangleBorder(
+//                   borderRadius: BorderRadius.circular(80))),
+//               backgroundColor:
+//                   const MaterialStatePropertyAll(Color(0xFF202832)),
+//               padding: const MaterialStatePropertyAll(
+//                 EdgeInsets.symmetric(vertical: 16, horizontal: 12),
+//               ),
+//             ),
+//             onPressed: onPressed,
+//             child: Center(
+//                 child: Text(
+//               'Get Started',
+//               style: Theme.of(context).textTheme.labelLarge,
+//             ))));
+//     ;
+//   }
+// }
