+++ lib/theme/theme.dart
@@ -1,7 +1,9 @@
+import 'package:dcy_wallet/theme/button.dart';
 import 'package:dcy_wallet/theme/typography.dart';
 import 'package:flutter/material.dart';
 
 final ThemeData theme = ThemeData(
   scaffoldBackgroundColor: const Color(0xFF080A0C),
-  textTheme: textTheme
+  textTheme: textTheme,
+  elevatedButtonTheme: buttonTheme
 );
